refactor: implement handler factory for getOne and getAll controllers
