+++ controllers/handlerFactory.js
@@ -1,4 +1,5 @@
-const { catchAsync, AppError } = require('./../utils');
+const { populate } = require('../models/reviewModel');
+const { catchAsync, AppError, APIFeatures } = require('./../utils');
 exports.deleteOne = (Model) =>
   catchAsync(async (req, res, next) => {
     // controller logic
@@ -40,6 +41,47 @@ exports.createOne = (Model) =>
 
     res.status(201).json({
       status: 'success',
+      data: {
+        data: doc,
+      },
+    });
+  });
+
+exports.getOne = (Model, popOptions) =>
+  catchAsync(async (req, res, next) => {
+    // controller logic
+    let query = Model.findById(req.params.id);
+    if (popOptions) query = query.populate(popOptions);
+    const doc = await query;
+
+    if (!doc) {
+      return next(new AppError('No document found with that ID', 404));
+    }
+    res.status(200).json({
+      status: 'success',
+      data: {
+        data: doc,
+      },
+    });
+  });
+
+exports.getAll = (Model) =>
+  catchAsync(async (req, res, next) => {
+    // To allow for nested GET reviews on tour
+    let filter = {};
+    if (req.params.tourId) filter = { tour: req.params.tourId };
+
+    // EXECUTE QUERY
+    const features = new APIFeatures(Model.find(filter), req.query)
+      .filter()
+      .sort()
+      .limiting();
+    await features.pagination();
+    const doc = await features.query;
+
+    res.status(200).json({
+      status: 'success',
+      results: doc.length,
       data: {
         data: doc,
       },
