+++ routes/reviewRoutes.js
@@ -15,6 +15,7 @@ router
   );
 router
   .route('/:id')
+  .get(reviewController.getReview)
   .patch(reviewController.updateReview)
   .delete(reviewController.deleteReview);
